requirements.txt
certifi
chardet
concurrent-log-handler
ConcurrentLogHandler
configparser
Django==2.2.28
django-filter==2.2.0
djangorestframework==3.11.2
ga4mp==1.1.1
gunicorn==19.9.0
idna
Markdown
pkg-resources
portalocker
pytz
requests
sqlparse
stripe==2.48.0
urllib3
